creating template to be filled
